Fix interning empty string error when locale is ""
This issue popped up with the i18n_backend_database plugin while
using a mysql database for storing locale settings. Somewhere
along the lines before reaching the around_generate method
I18n.locale is getting unset and wreaking havoc. I switched to
checking if the locale is also a blank string in addition to just
being nil, and then reset I18n.locale to the current locale, and
all is working again. I am not very familiar with rails routing
internals or with this plugin, so this might not be the best
solution, but it solved my problem.